+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
     ![image](https://user-images.githubusercontent.com/72517530/204557970-0ba31f3b-5209-4707-a2d0-55cc45509f8a.png)
 -   Please note that the functionality of downloading cc.json maps will no longer be supported from April 2023. Instead, use the Custom View feature to create map configurations. [#3079](https://github.com/MaibornWolff/codecharta/pull/3076)
 
+### Chore 👨‍💻 👩‍💻
+
+-   Add attribute descriptors to SoureCodeParser [#3166](https://github.com/MaibornWolff/codecharta/pull/3166)
+
 ## [1.112.1] - 2022-12-01
 
 ### Fixed 🐞